管理者チェックを追加

+++ src/controllers/DestroyServlet.java
@@ -2,12 +2,14 @@ package controllers;
 
 import java.io.IOException;
 
+import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import models.Member;
 import services.MemberService;
 
 /**
@@ -29,23 +31,34 @@ public class DestroyServlet extends HttpServlet {
      */
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
-        String _token = request.getParameter("_token");
-        if(_token != null && _token.equals(request.getSession().getId())) {
+//      管理者チェック
+      Member mm = (Member) request.getSession().getAttribute("login_member");
 
-            MemberService service = new MemberService();
+      if(mm.getAdminFlag() == true) {
 
-//            idを条件に会員データを論理削除する
-            service.destroy(Integer.parseInt(request.getParameter("id")));
+            String _token = request.getParameter("_token");
+            if(_token != null && _token.equals(request.getSession().getId())) {
 
-            service.close();
+                MemberService service = new MemberService();
 
-//          セッションに登録完了のフラッシュメッセージを設定
-          request.getSession().setAttribute("flush","退会が完了しました。");
+//                idを条件に会員データを論理削除する
+                service.destroy(Integer.parseInt(request.getParameter("id")));
 
-//            トップページにリダイレクト
-            response.sendRedirect(request.getContextPath() + "/indextop");
+                service.close();
 
+//              セッションに登録完了のフラッシュメッセージを設定
+                request.getSession().setAttribute("flush","退会が完了しました。");
+
+//              トップページにリダイレクト
+                response.sendRedirect(request.getContextPath() + "/indextop");
+
+            } else {
+
+//              エラー画面を表示
+                RequestDispatcher rd = request.getRequestDispatcher("/WEB-INF/views/error/unknown.jsp");
+                rd.forward(request, response);
+
+            }
         }
     }
-
 }
